+++ Sources/Extensions/CALayer+Relations.swift
@@ -1,65 +1,63 @@
 //
-//  UIView+Relations.swift
-//  Framezilla
-//
-//  Created by Nikita on 26/08/16.
-//  Copyright © 2016 Nikita. All rights reserved.
+//  Copyright © 2020 Rosberry. All rights reserved.
 //
 
-public extension UIView {
+import UIKit
+
+public extension CALayer {
 
     /// Width relation of current view.
 
     var nui_width: RelationView<SizeRelation> {
-        return RelationView(view: .view(self), relation: .width)
+        return RelationView(view: .layer(self), relation: .width)
     }
 
     /// Height relation of current view.
 
     var nui_height: RelationView<SizeRelation> {
-        return RelationView<SizeRelation>(view: .view(self), relation: .height)
+        return RelationView<SizeRelation>(view: .layer(self), relation: .height)
     }
 
     /// Left relation of current view.
 
     var nui_left: RelationView<HorizontalRelation> {
-        return RelationView<HorizontalRelation>(view: .view(self), relation: .left)
+        return RelationView<HorizontalRelation>(view: .layer(self), relation: .left)
     }
 
     /// Right relation of current view.
 
     var nui_right: RelationView<HorizontalRelation> {
-        return RelationView<HorizontalRelation>(view: .view(self), relation: .right)
+        return RelationView<HorizontalRelation>(view: .layer(self), relation: .right)
     }
 
     /// Top relation of current view.
 
     var nui_top: RelationView<VerticalRelation> {
-        return RelationView<VerticalRelation>(view: .view(self), relation: .top)
+        return RelationView<VerticalRelation>(view: .layer(self), relation: .top)
     }
 
     /// Bottom relation of current view.
 
     var nui_bottom: RelationView<VerticalRelation> {
-        return RelationView<VerticalRelation>(view: .view(self), relation: .bottom)
+        return RelationView<VerticalRelation>(view: .layer(self), relation: .bottom)
     }
 
     /// CenterX relation of current view.
 
     var nui_centerX: RelationView<HorizontalRelation> {
-        return RelationView<HorizontalRelation>(view: .view(self), relation: .centerX)
+        return RelationView<HorizontalRelation>(view: .layer(self), relation: .centerX)
     }
 
     /// CenterY relation of current view.
 
     var nui_centerY: RelationView<VerticalRelation> {
-        return RelationView<VerticalRelation>(view: .view(self), relation: .centerY)
+        return RelationView<VerticalRelation>(view: .layer(self), relation: .centerY)
     }
 
     /// Safe area
 
     @available(iOS 11.0, *)
     var nui_safeArea: SafeAreaRelationCollection {
-        return SafeAreaRelationCollection(view: .view(self))
+        return SafeAreaRelationCollection(view: .layer(self))
     }
 }
