+++ Sources/MakerHelper.swift
@@ -8,24 +8,9 @@
 
 import Foundation
 
-fileprivate extension UIView {
-
-    func contains(_ view: UIView) -> Bool {
-        if subviews.contains(view) {
-            return true
-        }
-        for subview in subviews {
-            if subview.contains(view) {
-                return true
-            }
-        }
-        return false
-    }
-}
-
 extension Maker {
 
-    func convertedValue(for type: RelationType, with view: UIView) -> CGFloat {
+    func convertedValue(for type: RelationType, with view: ViewType) -> CGFloat {
         var rect: CGRect {
             if let superview = self.view.superview, superview === view {
                 return CGRect(origin: .zero, size: superview.frame.size)
@@ -39,7 +24,11 @@ extension Maker {
         }
 
         var convertedRect = rect
-        if let superScrollView = self.view.superview as? UIScrollView, view is UIScrollView, superScrollView.contentSize != .zero {
+        if case let .view(superview) = self.view.superview,
+            case let .view(view) = view,
+            let superScrollView = superview as? UIScrollView,
+            view is UIScrollView,
+            superScrollView.contentSize != .zero {
             convertedRect.size.width = superScrollView.contentSize.width
             convertedRect.size.height = superScrollView.contentSize.height
         }
@@ -47,7 +36,7 @@ extension Maker {
         return value(for: type, with: view, in: convertedRect)
     }
 
-    func value(for type: RelationType, with view: UIView, in rect: CGRect) -> CGFloat {
+    func value(for type: RelationType, with view: ViewType, in rect: CGRect) -> CGFloat {
         switch type {
         case .top:
             return rect.minY
@@ -62,35 +51,19 @@ extension Maker {
         case .left:
             return rect.minX
         case .safeArea(.top):
-            if #available(iOS 11.0, *) {
-                return rect.minY + view.safeAreaInsets.top
-            }
-
-            return rect.minY
+            return rect.minY + view.safeAreaInsets.top
         case .safeArea(.left):
-            if #available(iOS 11.0, *) {
-                return rect.minX + view.safeAreaInsets.left
-            }
-
-            return rect.minX
+            return rect.minX + view.safeAreaInsets.left
         case .safeArea(.right):
-            if #available(iOS 11.0, *) {
-                return rect.maxX - view.safeAreaInsets.right
-            }
-
-            return rect.maxX
+            return rect.maxX - view.safeAreaInsets.right
         case .safeArea(.bottom):
-            if #available(iOS 11.0, *) {
-                return rect.maxY - view.safeAreaInsets.bottom
-            }
-
-            return rect.maxY
+            return rect.maxY - view.safeAreaInsets.bottom
         default:
             return 0
         }
     }
 
-    func relationSize(view: UIView, for type: RelationType) -> CGFloat {
+    func relationSize(view: ViewType, for type: RelationType) -> CGFloat {
         switch type {
         case .width:  return view.bounds.width
         case .height: return view.bounds.height
